공지사항목록, 상세, qna목록, faq목록

+++ src/main/java/com/petlogue/duopetbackend/admin/jpa/repository/QnaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QnaRepository extends JpaRepository<QnaEntity, Long> {
+public interface QnaRepository extends JpaRepository<QnaEntity, Integer> {
 
 
     Page<QnaEntity> findByContentType(String contentType, Pageable pageable);
